renamed profileStackHelpers to getProfileStackConfig
