+++ src/features/profile/api/useFeedback.ts
@@ -3,7 +3,7 @@ import { useMutation } from 'react-query'
 
 import { api } from 'api/api'
 import { PostFeedbackBody } from 'api/gen'
-import { getProfileStackConfig } from 'features/navigation/ProfileStackNavigator/profileStackHelpers'
+import { getProfileStackConfig } from 'features/navigation/ProfileStackNavigator/getProfileStackConfig'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { SNACK_BAR_TIME_OUT, useSnackBarContext } from 'ui/components/snackBar/SnackBarContext'
